[feature] 시장가 변동 집계 - Double 타입 -> Long으로 전부 변경 - Kafka 연동 - 일별 시장가 변동 집계 시 카프카 이벤트 발행(종료가, 거래량, 상품 uuid) - 스웨거 제거 - 포트번호 변경 8091
#2

+++ src/main/java/com/pieceofcake/batch_service/piece/dto/in/DailyFragmentAggregationDto.java
@@ -13,22 +13,22 @@ import java.time.LocalDate;
 public class DailyFragmentAggregationDto {
 
     private String pieceProductUuid;
-    private Double startingPrice;
-    private Double closingPrice;
-    private Double minimumPrice;
-    private Double maximumPrice;
-    private Double averagePrice;
+    private Long startingPrice;
+    private Long closingPrice;
+    private Long minimumPrice;
+    private Long maximumPrice;
+    private Long averagePrice;
     private Long tradeQuantity;
     private LocalDate date;
 
     @Builder
     public DailyFragmentAggregationDto(
             String pieceProductUuid,
-            Double startingPrice,
-            Double closingPrice,
-            Double minimumPrice,
-            Double maximumPrice,
-            Double averagePrice,
+            Long startingPrice,
+            Long closingPrice,
+            Long minimumPrice,
+            Long maximumPrice,
+            Long averagePrice,
             Long tradeQuantity,
             LocalDate date
     ){
