+++ src/main/java/com/pieceofcake/batch_service/piece/entity/DailyFragmentPriceAggregation.java
@@ -18,11 +18,11 @@ public class DailyFragmentPriceAggregation {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
     private String pieceProductUuid;
-    private Double startingPrice;
-    private Double closingPrice;
-    private Double maximumPrice;
-    private Double minimumPrice;
-    private Double averagePrice;
+    private Long startingPrice;
+    private Long closingPrice;
+    private Long maximumPrice;
+    private Long minimumPrice;
+    private Long averagePrice;
     private Long tradeQuantity;
     private LocalDate date;
 
@@ -30,11 +30,11 @@ public class DailyFragmentPriceAggregation {
     public DailyFragmentPriceAggregation(
             Long id,
             String pieceProductUuid,
-            Double startingPrice,
-            Double closingPrice,
-            Double maximumPrice,
-            Double minimumPrice,
-            Double averagePrice,
+            Long startingPrice,
+            Long closingPrice,
+            Long maximumPrice,
+            Long minimumPrice,
+            Long averagePrice,
             Long tradeQuantity,
             LocalDate date) {
         this.id = id;
